+++ src/components/ItemEditorPage/CenterPanel/AvatarColorDropdown/index.ts
@@ -1,0 +1,2 @@
+import AvatarColorDropdown from './AvatarColorDropdown'
+export default AvatarColorDropdown
